Move existing selector impl into `selector/simple`
